refactor: PartyDetailView Mock데이터 연결
